Refactor domain model with emphasis on inventory management
This commit involves major changes in the domain model to shift focus on inventory management. Deprecated classes such as Budget and Transaction are removed, FinancialGoal is renamed and repurposed as Inventory class to contain product and stock information. New classes are also introduced like Product, Supplier (which handles product supplying), and Email as a value object. This also includes the creation of new repository classes for these new model classes.